🧹 tidy up
#42
